Doc style (#113)
Adds doc string to all of the modules that have been switched from
private to public.

@christian-oreilly Options_mixin.py is currently left without docstring.
This file is generated by another and has a warning not to edit it
directly. Docstring shouldn't mess with anything, though. Should I add
the docstrings directly to that file, or should I find a way to add them
when generated with the generate_options_mixin.py file?

Closes #109

---------

Co-authored-by: Christian O'Reilly <christian.oreilly@gmail.com>