added frame4d and example (#119)
I created an example 4d volume browsing widget. This required adding the
"frame4D" parameter -- I just copied what was done for "colormap" and
"opacity". It uses nibabel to get the dimensions of the nifti file, but
this could be replaced by a method to get the dimensions of the
currently loaded niivue volume (this method surely exists but I haven't
found it yet).

---------

Co-authored-by: Christian O'Reilly <christian.oreilly@gmail.com>