Fix(colormap_label): More properly implement colormap_label
Create a colormap_label property that gets synced to the frontend. Technically this implementation is slightly incorrect since it accepts a colormap type as input, instead of a colormap_label, but it is a better implementation than using self.send(...) so it will suffice for now.

The reason why it's important to use synced traitlets instead of self.send (usually) is because they get updated on-volume/mesh-creation instead of having to wait for a custom-msg-event in the frontend.