refactor(opts)!: move NiiVue options to 'opts' attribute
Options can still be set directly on the NiiVue object, but this will be **removed in version 2.4.1**. Currently, doing so displays a DeprecationWarning. To migrate, set config options by modifying `nv.opts`.

The config options object is now a `traitlets.HasTraits` class where any change propagates a change into the parent `.opts` attribute. The reason for making this change is two-fold:

1. To be more consistent with NiiVue naming
2. To not confuse NiiVue options with other NiiVue attributes