Added indicator for backend-added objects
This protects against duplicate volumes getting added in case a volume gets added in the backend and then removed in the backend before the frontend has time to process it.

By default, before this change, the add_volume event in the frontend assumed that the volume was not added from the backend, but did not account for this add+remove spam situation. By adding "_py" to the end of the ID, we can clearly see if the object was added from the backend or not.

Also, used underscore instead of dash so that the ending can be easily removed if just the plain uuid is needed in the future.